requirements.txt
# API Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.4.0
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
pandas==2.2.0

# Vector Search
faiss-cpu==1.7.4

# Database
motor==3.3.2
pymongo==4.6.1

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0

# Logging
loguru==0.7.2

# Image Processing
opencv-python-headless==4.9.0.80

# Development
pytest==7.4.3
pytest-asyncio==0.23.4
black==24.1.1
flake8==7.0.0
